requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
python-dotenv==1.0.0
azure-identity==1.15.0
requests==2.31.0
gunicorn==21.2.0
